# Improve the performance of IterableSerie