--- a/pkg/metrics/iterable_series.go
+++ b/pkg/metrics/iterable_series.go
@@ -6,31 +6,33 @@
 package metrics
 
 import (
+	"context"
 	"errors"
 	"sync/atomic"
 
+	"github.com/DataDog/datadog-agent/pkg/util"
 	jsoniter "github.com/json-iterator/go"
 )
 
 // IterableSeries represents an iterable collection of Serie.
 // Serie can be appended to IterableSeries while IterableSeries is serialized
 type IterableSeries struct {
-	c                   chan *Serie
-	receiverStoppedChan chan struct{}
-	callback            func(*Serie)
-	current             *Serie
-	count               uint64
+	c        *util.BufferedChan
+	cancel   context.CancelFunc
+	callback func(*Serie)
+	current  *Serie
+	count    uint64
 }
 
 // NewIterableSeries creates a new instance of *IterableSeries
 // `callback` is called each time `Append` is called.
-// `chanSize` is the internal channel buffer size
-func NewIterableSeries(callback func(*Serie), chanSize int) *IterableSeries {
+func NewIterableSeries(callback func(*Serie), chanSize int, bufferSize int) *IterableSeries {
+	ctx, cancel := context.WithCancel(context.Background())
 	return &IterableSeries{
-		c:                   make(chan *Serie, chanSize),
-		receiverStoppedChan: make(chan struct{}),
-		callback:            callback,
-		current:             nil,
+		c:        util.NewBufferedChan(ctx, chanSize, bufferSize),
+		cancel:   cancel,
+		callback: callback,
+		current:  nil,
 	}
 }
 
@@ -38,12 +40,7 @@ func NewIterableSeries(callback func(*Serie), chanSize int) *IterableSeries {
 func (series *IterableSeries) Append(serie *Serie) {
 	series.callback(serie)
 	atomic.AddUint64(&series.count, 1)
-	select {
-	case series.c <- serie:
-
-	// Make sure `Append` doesn't block. See `IterationStopped()`.
-	case <-series.receiverStoppedChan:
-	}
+	series.c.Put(serie)
 }
 
 // SeriesCount returns the number of series appended with `IterableSeries.Append`.
@@ -53,7 +50,7 @@ func (series *IterableSeries) SeriesCount() uint64 {
 
 // SenderStopped must be called when sender stop calling Append.
 func (series *IterableSeries) SenderStopped() {
-	close(series.c)
+	series.c.Close()
 }
 
 // IterationStopped must be called when the receiver stops calling `MoveNext`.
@@ -61,7 +58,7 @@ func (series *IterableSeries) SenderStopped() {
 // end of the iteration because of an error and so blocks the sender forever
 // as no goroutine read the channel.
 func (series *IterableSeries) IterationStopped() {
-	close(series.receiverStoppedChan)
+	series.cancel()
 }
 
 // WriteHeader writes the payload header for this type
@@ -76,25 +73,31 @@ func (series *IterableSeries) WriteFooter(stream *jsoniter.Stream) error {
 
 // WriteCurrentItem writes the json representation of an item
 func (series *IterableSeries) WriteCurrentItem(stream *jsoniter.Stream) error {
-	if series.current == nil {
+	current := series.Current()
+	if current == nil {
 		return errors.New("nil serie")
 	}
-	return writeItem(stream, series.current)
+	return writeItem(stream, current)
 }
 
 // DescribeCurrentItem returns a text description for logs
 func (series *IterableSeries) DescribeCurrentItem() string {
-	if series.current == nil {
+	current := series.Current()
+	if current == nil {
 		return "nil serie"
 	}
-	return describeItem(series.current)
+	return describeItem(current)
 }
 
 // MoveNext advances to the next element.
 // Returns false for the end of the iteration.
 func (series *IterableSeries) MoveNext() bool {
-	var ok bool
-	series.current, ok = <-series.c
+	v, ok := series.c.Get()
+	if v != nil {
+		series.current = v.(*Serie)
+	} else {
+		series.current = nil
+	}
 	return ok
 }