--- a/pkg/metrics/iterable_series.go
+++ b/pkg/metrics/iterable_series.go
@@ -10,14 +10,15 @@ import (
 	"errors"
 	"sync/atomic"
 
-	"github.com/DataDog/datadog-agent/pkg/util"
 	jsoniter "github.com/json-iterator/go"
+
+	"github.com/DataDog/datadog-agent/pkg/util"
 )
 
 // IterableSeries represents an iterable collection of Serie.
 // Serie can be appended to IterableSeries while IterableSeries is serialized
 type IterableSeries struct {
-	c        *util.BufferedChan
+	ch       *util.BufferedChan
 	cancel   context.CancelFunc
 	callback func(*Serie)
 	current  *Serie
@@ -29,7 +30,7 @@ type IterableSeries struct {
 func NewIterableSeries(callback func(*Serie), chanSize int, bufferSize int) *IterableSeries {
 	ctx, cancel := context.WithCancel(context.Background())
 	return &IterableSeries{
-		c:        util.NewBufferedChan(ctx, chanSize, bufferSize),
+		ch:       util.NewBufferedChan(ctx, chanSize, bufferSize),
 		cancel:   cancel,
 		callback: callback,
 		current:  nil,
@@ -40,7 +41,7 @@ func NewIterableSeries(callback func(*Serie), chanSize int, bufferSize int) *Ite
 func (series *IterableSeries) Append(serie *Serie) {
 	series.callback(serie)
 	atomic.AddUint64(&series.count, 1)
-	series.c.Put(serie)
+	series.ch.Put(serie)
 }
 
 // SeriesCount returns the number of series appended with `IterableSeries.Append`.
@@ -50,7 +51,7 @@ func (series *IterableSeries) SeriesCount() uint64 {
 
 // SenderStopped must be called when sender stop calling Append.
 func (series *IterableSeries) SenderStopped() {
-	series.c.Close()
+	series.ch.Close()
 }
 
 // IterationStopped must be called when the receiver stops calling `MoveNext`.
@@ -92,7 +93,7 @@ func (series *IterableSeries) DescribeCurrentItem() string {
 // MoveNext advances to the next element.
 // Returns false for the end of the iteration.
 func (series *IterableSeries) MoveNext() bool {
-	v, ok := series.c.Get()
+	v, ok := series.ch.Get()
 	if v != nil {
 		series.current = v.(*Serie)
 	} else {