--- a/pkg/util/buffered_chan.go
+++ b/pkg/util/buffered_chan.go
@@ -11,7 +11,9 @@ import (
 )
 
 // BufferedChan behaves like a `chan []interface{}` (See thread safety for restrictions), but is
-// most efficient as it uses internally a channel of []interface{}.
+// most efficient as it uses internally a channel of []interface{}. This reduces the number of reads
+// and writes to the channel. Instead of having one write and one read for each value for a regular channel,
+// there are one write and one read for each `bufferSize` value.
 // Thread safety:
 // 	- `BufferedChan.Put` cannot be called concurrently.
 // 	- `BufferedChan.Get` cannot be called concurrently.