@ogaca-dd 2022-01-05
@djmitche 

> What is the advantage over "just" a channel here? I see that the benchmarks show the overal serialization is faster, but why?

Inserting into a slice is faster than inserting into a channel. The drawback is that two goroutines cannot push to `BufferedChan` (This is the same for reading).


> I'm also curious if the use of interface{} here has a cost, either in terms of memory overhead (an interface includes TypeInfo) or CPU (that TypeInfo must be checked on cast). Would it be more efficient to specialize this channel to a particular datatype, perhaps through code generation?

The initial implementation used `*Serie` and not interface. I did a small benchmark that shows the difference in term of CPU and allocations was minor (at least in my case).

Note: This class is a good candidate for a refactoring when generic will be available in Go :).